requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly.express
pydeck
streamlit-aggrid